extract alert destinations service
