+++ apps/api/src/platform/describe-cause.ts
@@ -1,0 +1,11 @@
+/** Render an unknown nested cause for logs and serialized persistence errors. */
+export const describeCause = (cause: unknown): string | undefined => {
+	if (cause == null) return undefined
+	if (cause instanceof Error) return cause.stack ?? cause.message
+	if (typeof cause === "string") return cause
+	try {
+		return JSON.stringify(cause)
+	} catch {
+		return String(cause)
+	}
+}
