+++ apps/api/src/routes/v2/alerts.http.test.ts
@@ -17,6 +17,7 @@ import { ApiKeysService } from "@/services/org/ApiKeysService"
 import { AuthService } from "@/services/auth/AuthService"
 import { DashboardPersistenceService } from "@/services/dashboards/DashboardPersistenceService"
 import { AlertRuntime, AlertsService } from "@/services/alerts/AlertsService"
+import { AlertDestinationsService } from "@/services/alerts/AlertDestinationsService"
 import { HazelOAuthService } from "@/services/auth/HazelOAuthService"
 import { OrgClickHouseSettingsService } from "@/services/org/OrgClickHouseSettingsService"
 import { OrgMembersService } from "@/services/org/OrgMembersService"
@@ -100,6 +101,11 @@ const makeHarness = (warehouseService: WarehouseQueryServiceShape = warehouseStu
 	const orgChSettingsLive = OrgClickHouseSettingsService.layer.pipe(
 		Layer.provide(Layer.mergeAll(envLive, testDb.layer)),
 	)
+	const alertDestinationsLive = AlertDestinationsService.layer.pipe(
+		Layer.provide(
+			Layer.mergeAll(envLive, testDb.layer, runtimeLive, hazelOAuthLive, emailLive, orgMembersLive),
+		),
+	)
 	const alertsLive = AlertsService.layer.pipe(
 		Layer.provide(
 			Layer.mergeAll(
@@ -113,6 +119,7 @@ const makeHarness = (warehouseService: WarehouseQueryServiceShape = warehouseStu
 				orgMembersLive,
 				orgChSettingsLive,
 				investigationsLive,
+				alertDestinationsLive,
 			),
 		),
 	)
@@ -120,6 +127,7 @@ const makeHarness = (warehouseService: WarehouseQueryServiceShape = warehouseStu
 		ApiKeysService.layer,
 		AuthService.layer,
 		DashboardPersistenceService.layer,
+		alertDestinationsLive,
 		alertsLive,
 	).pipe(Layer.provideMerge(Layer.mergeAll(envLive, testDb.layer)))
